+++ TMSHomework-Lesson3/TMSHomework-Lesson3/Task4 - callRecursedFunction.swift
@@ -1,0 +1,14 @@
+//
+//  Task4 - callRecursedFunction.swift
+//  TMSHomework-Lesson3
+//
+//  Created by Наталья Мазур on 2.12.23.
+//
+
+import Foundation
+
+//    TODO: 4. Циклический вызов функций - поломать приложение
+
+func callRecursedFunction() {
+    callRecursedFunction()
+}
