+++ TMSHomework-Lesson3/TMSHomework-Lesson3/Task6 - calculateClosures.swift
@@ -1,0 +1,10 @@
+//
+//  Task6 - calculateClosures.swift
+//  TMSHomework-Lesson3
+//
+//  Created by Наталья Мазур on 2.12.23.
+//
+
+import Foundation
+
+//    TODO: 6. Универсальные функции сложения, вычитания, умножения и деления через замыкание
