+++ TMSHomework-Lesson3/TMSHomework-Lesson3/Task8 - checkFlowerWatering.swift
@@ -1,0 +1,10 @@
+//
+//  Task8 - checkFlowerWatering.swift
+//  TMSHomework-Lesson3
+//
+//  Created by Наталья Мазур on 2.12.23.
+//
+
+import Foundation
+
+//    TODO: 8. Замыкание в котором ты проверяешь полил ли ты цветок, если нет то напоминание полить цветок
