+++ TMSHomework-Lesson3/TMSHomework-Lesson3/Task9 - runClosureInsideClosure.swift
@@ -1,0 +1,10 @@
+//
+//  Task9 - runClosureInsideClosure.swift
+//  TMSHomework-Lesson3
+//
+//  Created by Наталья Мазур on 2.12.23.
+//
+
+import Foundation
+
+//    TODO: 9. Выполнить любое замыкание внутри замыкания по очереди/ замыкание внутри любой функции
